move section code check to SectionsAsStudentTableTest
